+++ client/src/components/Header/headerStyling.css
@@ -310,7 +310,7 @@
     }
 
     .my_journey_button {
-        margin-top: 10rem;
+        margin-top: 12rem;
     }
     
     .scroll-btn {
